requirements.txt
anthropic==0.39.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
beautifulsoup4==4.12.3
redis==5.2.0
python-dotenv==1.0.1
pydantic==2.10.2
